backend/requirements.txt
# backend/requirements.txt

fastapi==0.115.0
uvicorn==0.30.1
sqlmodel==0.0.22
pydantic==2.8.2
python-dotenv==1.0.1
pytest==8.3.1
pandas==2.2.3
